[refactor] views.pyに記述していたViewクラスをviewsディレクトリ内に分割
