+++ match_service/match_app/urls.py
@@ -1,10 +1,8 @@
 from django.urls import path
 
-from .views import (
-    MatchHistoryView,
-    MatchView,
-    health_check,
-)
+from match_app.views.health_check import health_check
+from match_app.views.match_view import MatchView
+from match_app.views.match_history_view import MatchHistoryView
 
 urlpatterns = [
     path("matches/", MatchView.as_view(), name="matches"),
